Implements "context bubbling" as defunkt said it was meant to be in the mustache spec (https://github.com/janl/mustache.js/issues#issue/65).
It seems this behavior was desired in mustache.js as [examples/complex.js](http://agib.me/ha2dUq) uses pseudo bubbling; i.e., the main context is always accessible sub-contexts. However, all contexts in between a sub-context and the main context were inaccessible... That's where this patch comes in.

Now, as create_context() gets called for mustache sections it gets passed the current context to be saved as the "parent" context for the newly created sub-context. This way, find() can always fallback the chain of parent contexts (bubbling). This eliminates the need for the pseudo bubbling mentioned above and implemented on lines 221-222. Also, when falling back, find() always sends the original sub-context as a third parameter so that functions get called with the appropriate context. Finally, we need to ensure that we don't allow infinite loops -- I found this happening in [examples/partial_recursion.js](http://agib.me/gDOEDZ). We just need to mark visited sub-contexts.