Adds note about needing rspec to run rake commands.
I was getting the following error when running `rake commonjs`:

```
rake aborted!
no such file to load -- rspec/core/rake_task

(See full trace by running task with --trace)
```

This was with a default OS X Lion install. The issue was simply that
rspec was not installed, however, it may not be obvious at first.