Remove ability to use a "streaming" function
The "send" function from the 0.4 API doesn't actually stream the template
so much as it just passes chunks of the rendered template through to the
callback as they are rendered. The callbacks are fired synchronously either
way, and using a "send" function just adds extra overhead and makes things
slower.