Make high-level API same as Renderer API
Also, allow partials to be passed to the function that is returned
for pre-compiled templates. This makes the API more consistent all
around, but comes with the caveat that renderers are much more
aggressive at caching. Now, any time a template or partial is
compiled with the same name as a previously compiled template the
cached version is used.

Fixes #240