Don't evaluate lambdas that return falsy values as an empty string.
Issue: If I create a Mustache lambda to return numbers and I expect 0 to be of
those numbers, I'll instead only receive an empty string for 0 because it
evaluates to falsy.

Fix: Check the result of the lambda to be undefined or null. Only then should
it return the empty string.