Fixed bug where nested partials would output [Object object] for empty strings.
When I used a nested partial template with an array that contained an empty string, Mustache would incorrectly output [Object object] in place of that string. I found that the bug was due to conditional statement that replaced all "falsy" values with an object, so I replaced it with an explicit check for undefined values.