Fix browser tests by running them on Node.js 6
For a long time we've struggled getting the browser test suite to run
on sauce labs. Those tests were running on Node.js 4 which is getting
quite old.

In a trial-n-error effort, I tried adding Node.js 6 as one of the
versions to run on Travis CI, and at the same time ensure the browsers
tests were run on that version instead of Node.js 4 -- voila the suite
passed right away.

It therefore seemed like the issues we were seeing had to do with us
running those tests via Node.js 4.