Cut official support for Node.js 0.8 by not running tests on it anymore
Primarily because it's very painful to install 3rd party packages
in the test phase of our CI runs, because many of them requires at least
Node.js 0.10 or even higher.

Therefore getting the test suite to run when using Node.js 0.8 is
far from trivial, and the benefit of ensure Node.js 0.8 support these
days is not obvious because it should really not be used knowing that
it has reached end-of-life a long time ago.

Worth noting there's no breaking changes in this commit to
mustache.js's source code, we just don't want to struggle getting our
CI to care about it. In other words, it should still be possible to use
Node.js 0.8, but we won't care much making sure it keeps working going
forward.