Bugfix for wrong functions output in partials with indentation
This small change fixes the output of functions used in partials with
indentation. Bug reports has shown that the functions output is shifted
with the amount of indentation the partial has.

The bug itself is best illustrated in the tests added in 621ae80652b0ef363f5c8f2aab4c5e885b0d953e.

Closes https://github.com/janl/mustache.js/issues/712