Bump eslint v2 -> v6
Primarily because it's been ages since we bumped eslint and as a precursor
to allowing some future test files to be written in a more modern type
of JavaScript. Especially think about tests written using puppeteer.js
which is heavily `Promise` based, where the async-await syntax makes
it soooo less painful to write.

Silly not being able to use that because of an outdated eslint version
doesn't understand that syntax construct at all.