Add CI test to verify build output is in sync with source (.js vs .mjs)
With the introduction of a build step, where we take the source in .mjs
and build it into .js for non-ES Module systems, it's very important
that we remember to keep those two in sync.

That's what the CI job created in these changes ensure;

1. Run the build script
2. See if there are any pending git changes as a result

As long as there are no pending git changes, we're all fine and dandy,
or else we need to run the `npm run build` and commit the changes.