Update usage examples to not include jQuery
Historically jQuery was seen as an absolute minimal to create
anything with JavaScript. That time has past now that relatively
modern browsers has excellent support for a lot of the things
jQuery helped us with -- at least with the trivial examples shown
in our README.

Therefore removing jQuery usage from our examples as that should
not be a necessary dependency in these examples.