Only keep Node.js 8 with browser usage tests on Travis CI
Removes running Node.js 10 & 12 from Travis CI as we do that with
GitHub Actions as well.

The only reason we still keep Node.js 8 running on Travis CI is due to
the browser usage testing we've setup, with encrypted secrets and all.
Ideally we'd move those browser tests to run via GitHub Actions for
consistency with the rest of our tests, but postponing that work
for now as it's not critical.