Remove nuget.org spec from the repository
None of the recent maintainers has remembered trying to keep the
nuget.org package up to date.

As of writing this, it's over 7 years since the last mustache.js
version got published on nuget (v0.7.2).

Adding to that, there has been no questions asked from the nuget
community about outdated versions.

With the above in mind, we'll remove the nuget spec from our repo
since it only adds to confusion at this point.

Refs https://www.nuget.org/packages/mustache.js/